Split the batching channel module into facade, worker engine, and grouped test files
