+++ crates/burn-lm-inference/src/channels/batching/tests/mod.rs
@@ -1,0 +1,7 @@
+mod fakes;
+
+mod cancellation;
+mod concurrency;
+mod failure;
+mod lifecycle;
+mod stress;
